+++ scm-webapp/src/main/java/sonia/scm/api/v2/resources/BranchDto.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-    
+
 package sonia.scm.api.v2.resources;
 
 import de.otto.edison.hal.Embedded;
@@ -31,17 +31,23 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.hibernate.validator.constraints.Length;
+
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.Pattern;
 
-@Getter @Setter @NoArgsConstructor
+@Getter
+@Setter
+@NoArgsConstructor
+@SuppressWarnings("java:S2160") // we do not need this for dto
 public class BranchDto extends HalRepresentation {
 
   private static final String VALID_CHARACTERS_AT_START_AND_END = "\\w-,;\\]{}@&+=$#`|<>";
   private static final String VALID_CHARACTERS = VALID_CHARACTERS_AT_START_AND_END + "/.";
   static final String VALID_BRANCH_NAMES = "[" + VALID_CHARACTERS_AT_START_AND_END + "]([" + VALID_CHARACTERS + "]*[" + VALID_CHARACTERS_AT_START_AND_END + "])?";
 
-  @NotEmpty @Length(min = 1, max=100) @Pattern(regexp = VALID_BRANCH_NAMES)
+  @NotEmpty
+  @Length(min = 1, max = 100)
+  @Pattern(regexp = VALID_BRANCH_NAMES)
   private String name;
   private String revision;
   private boolean defaultBranch;